Moved main out of vm.c and adjusted make test in makefile
