+++ stackVM/makefile
@@ -3,8 +3,8 @@ CFLAGS=-Wall -Wextra -std=c11 -lm
 SOURCES=src/*.c
 TESTS=tests/*.c
 TARGET=bolt
-TESTSRC=src/dataconstant.c src/frame.c src/builtin.c src/impl_builtin.c src/stringvector.c
-TESTBUILD=tests/runtests
+TEST_SOURCES=src/[!app]*.c
+TESTBUILD=tests/bolt-tests
 
 default: clean $(TARGET)
 
@@ -12,7 +12,7 @@ $(TARGET): $(SOURCES-.c=.o)
 	$(CC) $(CFLAGS) -o $(TARGET) $(SOURCES)
 
 $(TESTBUILD): $(TESTS- .c=.o)
-	$(CC) $(CFLAGS) $< -lcriterion -o $(TESTBUILD) $(TESTS) $(TESTSRC)
+	$(CC) $(CFLAGS) $< -lcriterion -o $(TESTBUILD) $(TESTS) $(TEST_SOURCES)
 
 test: $(TESTBUILD)
 	@echo "Setting up tests..."
@@ -22,3 +22,4 @@ test: $(TESTBUILD)
 
 clean:
 	@rm -f $(TARGET)
+	@rm -f $(TESTBUILD)
